联动后端拿到主页列表数据

+++ app/src/main/java/com/hzr/kemo/model/DrinkListEntity.kt
@@ -6,7 +6,7 @@ import kotlinx.serialization.Serializable
 data class DrinkListEntity (
     val id: Int,
     val category: String,
-    val bgColor: Int,
+    val bgColor: Long,
     val imgSrc:String,
     val title: String,
     val desc: String,
